requirements.txt
torch~=2.2.2
numpy~=1.26.4
easydict~=1.13
neptune~=1.10.2
python-dotenv~=1.0.1
pandas~=2.2.2
evaluate~=0.4.2
tqdm~=4.66.4
transformers~=4.41.2
pillow~=10.4.0
youtokentome~=1.0.6